Test the ability to export an SVG
It's surprisingly hard to convert and SVG into a PNG in JavaScript.
While that's possible, it seems easier to draw directly to the canvas, with no SVG elements.

I'm freezing the current state before I rip out the SVG.

Added the ffmpeg.js package.
I added the simplest of test code for this.